uir/requirements.txt
Flask==1.0.2
pymongo==3.7.2
PyYAML==3.12
urllib3==1.2
